Move CopyBidirectional to stream_tools
