+++ penguin-mux/src/stream_tools/mod.rs
@@ -1,0 +1,5 @@
+//! Tools for working with [`crate::MuxStream`]s.
+//
+// SPDX-License-Identifier: Apache-2.0 OR GPL-3.0-or-later
+#[cfg(feature = "std")]
+pub mod copy_bidirectional;
